backend/pyproject.toml
[project]
name = "izumi-studio-backend"
version = "0.1.0"
description = "Backend service for Izumi Studio"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "pypdf>=6.0.0,<7.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0,<25.0.0",
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "ruff>=0.8.0,<1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ["py311"]
